first working version

+++ iscan_docker.sh
@@ -1,11 +1,33 @@
 #!/bin/bash
 
+
+##########################################################################
+#
+#  Author : Ton Schoots
+#  version: 0.1
+#
+#  This script scans all the containers on a docker host
+#  meaning a host able to perform docker commands.
+#
+#  dependencies:
+#       Black Duck hub iscan
+#       tar
+#       grep
+#       awk
+#       sort
+#       docker
+#
+#  Status: 29 - 9 - 2015
+#  this script is capable to do a i scan of all the images on the machine where it's run
+#  it does not however remove the BOM if an image is not there anymore.
+##########################################################################
+
 _ISCAN_CLIENT="/Users/tschoots/tmp/scanner/scan.cli-2.2.1.1/bin/scan.cli.sh"
 _HOSTNAME=$(hostname)
 
 usage() {
   cat <<EEOPTS
-    $(basename $0) -h <hub_hostname> -p <port> -s <scheme [http/https]> -u <username> -p <password> -i <image:tag>
+    $(basename $0) -h <hub_hostname> -p <port> -s <scheme [http/https]> -u <username> -w <password> -i <image:tag>
     Note all parameters are mandatory, except for the -i parameter which gives the opertunity to scan one image!
 EEOPTS
     exit 1
@@ -20,11 +42,16 @@ if [ "$version" != "Server:" ]; then
    exit
 fi
 
+if [ "$#" -ne "10" ] && [ "$#" -ne "12" ]; then
+   usage
+fi
+
 #check input parameters
-while getopts ":h:p:s:u:p:" opt; do
+while getopts "h:p:s:u:w:i:" opt; do
+   echo $opt
    case $opt in
       h)
-       host=$OPTARGA
+       host=$OPTARG
        ;;
       p)
        port=$OPTARG
@@ -35,8 +62,11 @@ while getopts ":h:p:s:u:p:" opt; do
       u)
        username=$OPTARG
        ;;
-      p)
+      w)
        password=$OPTARG
+       ;;
+      i)
+       image=$OPTARG
        ;;
       \?)
        usage
@@ -44,20 +74,26 @@ while getopts ":h:p:s:u:p:" opt; do
    esac
 done
 
-mkdir tmp
-cd tmp
-# get all the images sorted on size so the smallest will go first to avoid long waiting time in case of wrong parameters.
-declare -a images=($(docker images | grep -v "REPOSITORY\|<none>" | sort -k 5 -r | awk '{print $1":"$2}'))
+declare -a images=($(docker images $image | grep -v "REPOSITORY\|<none>" | sort -k 5 -r | awk '{print $1":"$2}'))
+if [ "${#images[@]}" -eq "0" ];then 
+   echo "no images found"
+   exit
+fi
 for img in "${images[@]}"
 do
+  mkdir tmp
+  cd tmp
   docker save -o dump.tar $img
-  tar -xvf dump.tar
+  tar -xf dump.tar
   rm -rf dump.tar
-  find . -name "*.tar" -exec tar -xvf {} \;
+  find . -name "*.tar" -exec tar -xf {} \;
   find . -name "*.tar" -exec rm -rf {} \;
-  result=$($_ISCAN_CLIENT $@ .)
+  cmd="$_ISCAN_CLIENT --host $host --port $port --scheme $scheme --project $(hostname) --release $img --username $username --password $password  -v ."
+  echo $cmd
+  result=$($cmd)
+  cd ..
+  chmod -R 777 tmp
+  rm -rf tmp
   echo $img
 done
 
-cmd="$_ISCAN_CLIENT --dryRunReadFile prr  $@"
-$cmd
